# feat: one `params` per a handler (optimized for RegExpRouter)