--- a/deno_dist/compose.ts
+++ b/deno_dist/compose.ts
@@ -1,4 +1,5 @@
 import { Context } from './context.ts'
+import type { ParamIndexMap, Params } from './router.ts'
 import type { Env, NotFoundHandler, ErrorHandler } from './types.ts'
 
 interface ComposeContext {
@@ -8,11 +9,10 @@ interface ComposeContext {
 
 // Based on the code in the MIT licensed `koa-compose` package.
 export const compose = <C extends ComposeContext, E extends Env = Env>(
-  middleware: Function[],
+  middleware: [Function, ParamIndexMap | Params][],
   onError?: ErrorHandler<E>,
   onNotFound?: NotFoundHandler<E>
 ) => {
-  const middlewareLength = middleware.length
   return (context: C, next?: Function) => {
     let index = -1
     return dispatch(0)
@@ -21,12 +21,20 @@ export const compose = <C extends ComposeContext, E extends Env = Env>(
       if (i <= index) {
         throw new Error('next() called multiple times')
       }
-      let handler = middleware[i]
       index = i
-      if (i === middlewareLength && next) handler = next
 
       let res
       let isError = false
+      let handler
+
+      if (middleware[i]) {
+        handler = middleware[i][0]
+        if (context instanceof Context) {
+          context.req.setParams(middleware[i][1])
+        }
+      } else {
+        handler = (i === middleware.length && next) || undefined
+      }
 
       if (!handler) {
         if (context instanceof Context && context.finalized === false && onNotFound) {