--- a/deno_dist/router/trie-router/router.ts
+++ b/deno_dist/router/trie-router/router.ts
@@ -22,7 +22,7 @@ export class TrieRouter<T> implements Router<T> {
     this.node.insert(method, path, handler)
   }
 
-  match(method: string, path: string): Result<T> | null {
+  match(method: string, path: string): Result<T> {
     return this.node.search(method, path)
   }
 }