--- a/src/context.ts
+++ b/src/context.ts
@@ -154,10 +154,6 @@ export class Context<
     this.finalized = true
   }
 
-  setRequestParams = (params: Record<string, string>) => {
-    this.req.setParams(params)
-  }
-
   /**
    * @experimental
    * `c.render()` is an experimental feature.