@yusukebe 2023-10-13
Hi @usualoma,

I have taken the time to think about this and I find it to be an interesting approach!

> However, since the contribution to speed by making it an Array is limited and the code is less readable, some may argue that the format using Objects is better in this regard.

Indeed, but I don't think it's significantly less readable.

So, how about we go this way!? If so, can you complete the PR? It's a hard task to write the code for other routers, but I'd like you to do it.

Thanks!