@usualoma 2023-10-13
@yusukebe OK, I will try to implement this specification in other routers.