@usualoma 2023-10-13
@yusukebe 
By the way, TrieRouter no longer checks for "Duplicate param name" at all, so it no longer makes an error if multiple cases of the same name exist in a single path.

Before
https://github.com/honojs/hono/blob/main/src/router/trie-router/node.test.ts#L435-L440

After
https://github.com/usualoma/hono/blob/48127094af314dad390907f727d1353ff322c4f1/src/router/trie-router/node.test.ts#L451-L458

What do you think about this point with other routers?

In my opinion, it is better to "not check Duplicate param name at all, even on other routers".
The reason for this is that there is no need for the framework to check it, since it is unlikely that there are multiple instances of the same name in a single path.
Instead, stopping checking and making the code shorter would be more useful.