@usualoma 2023-10-13
@yusukebe 
Oh, sorry, I didn't explain myself well enough.

RegExpRouter still makes an error for duplicates like `/:id/:id` in this branch, but I'm talking about whether or not to stop doing that check.
https://github.com/usualoma/hono/blob/48127094af314dad390907f727d1353ff322c4f1/src/router/reg-exp-router/node.ts#L100-L102