@yusukebe 2023-10-13
@usualoma

> What do you think about this point with other routers?
> In my opinion, it is better to "not check Duplicate param name at all, even on other routers".

I agree. We don't need to check it. And I've implemented it to make `id` be `123` not `456` in the test of TrieRouter:

```ts
it('self', () => {
  const node = new Node()
  node.insert('get', '/:id/:id', 'foo')
  const [res] = node.search('get', '/123/456')
  expect(res.length).toBe(1)
  expect(res[0][0]).toBe('foo')
  expect(res[0][1]['id']).toBe('123')
})
```

I think it's sufficient to have a rule that if the path has the same param name, the first value will be used.