@yusukebe 2023-10-13
Should it throw `UnsupportedPathError`?

> https://github.com/usualoma/hono/blob/48127094af314dad390907f727d1353ff322c4f1/src/router/reg-exp-router/node.ts#L100-L102