@yusukebe 2023-10-13
> Should it throw `UnsupportedPathError`?

Ah, if we could throw an error in the insert phase, that would be best.