@usualoma 2023-10-15
Done.

@yusukebe Review, please.

### For `/:id/:id`

* LinearRouter, RegExpRouter, TrieRouter : If the path is `/123/456`, `id` be `123` not `456`.
* PatternRouter : `throw new UnsupportedPathError()` in the insert phase