@yusukebe 2023-10-15
Hi @usualoma,

Thank you very much for the great work. It looks good!

I've added some tests and merged the `next` branch. I think there's no problem, so I'll merge it tomorrow morning.